refactor(semantic-graph): address review — extract id helpers, fix imports
- Move _slug_id / _CLEAN_ID_RE into backend/semantic_graph/id_utils.py so
  postprocessor no longer imports them from the heavy sympy_translator module
  (avoids the cross-module private import + circular-dependency smell flagged
  in review). Both translator and postprocessor import from id_utils.
- signature.py imports SemanticGraph/SemanticGraphNode from the public
  backend.model re-export rather than backend.model.semantic_graph.

No behavior change; full suite green (3483 passed).
