+++ backend/semantic_graph/sympy_translator.py
@@ -26,6 +26,7 @@ from sympy.physics.quantum import InnerProduct, OuterProduct
 
 from backend.model.semantic_graph import SemanticGraph
 
+from .id_utils import _slug_id
 from .preprocessor import LaTeXPreprocessor, strip_trailing_spacing
 from .constants import (
     KNOWN_VARIABLES,
@@ -4096,38 +4097,11 @@ def _resolve_xi_node_ids(graph: SemanticGraph) -> None:
             edge.to = rename_map[edge.to]
 
 
-# A node id is an internal wiring key, never a display string (display always
-# comes from ``latex`` / ``subexpr`` / ``label``). A clean id is a plain
-# identifier — letters, digits, underscores. The builder mints ids in this
-# form at node-creation time (see ``GraphBuilder._mint_symbol_id``); this regex
-# is the invariant the rest of the pipeline can rely on.
-_CLEAN_ID_RE = re.compile(r"^[A-Za-z0-9_]+$")
-
 # Synthetic collapse sentinels (multichar subscripts, ``\text{…}``, compound
 # identifiers, braket inner products) — the only override keys safe to
 # substitute back into a symbol's display latex. A leading ``cN_`` is the
 # per-clause prefix used in comma-separated statements.
 _PLACEHOLDER_KEY_RE = re.compile(r"^(?:c\d+_)?(?:Xi|Theta|Phi)_\{\d+\}$")
-
-
-def _slug_id(raw: str) -> str:
-    """Reduce a (possibly LaTeX-bearing) string to a clean ``[A-Za-z0-9_]``
-    identifier suitable for a node id. LaTeX/display content lives in the
-    node's ``latex`` / ``subexpr`` — the id only needs to be a stable,
-    collision-free wiring key.
-
-        ``V_{\\text{exit}}`` → ``V_exit``      ``\\gamma_{\\alpha}`` → ``gamma_alpha``
-        ``I_{sp}``           → ``I_sp``        ``\\epsilon_{0}``     → ``epsilon_0``
-        ``\\Delta \\gamma``  → ``Delta_gamma``
-    """
-    if _CLEAN_ID_RE.fullmatch(raw):
-        return raw  # already clean (plain symbols, operator ids) — leave as-is
-    s = re.sub(r"\\text\{([^{}]*)\}", r"\1", raw)   # \text{exit} → exit
-    s = s.replace("\\", "")                          # \gamma → gamma
-    s = s.replace("{", "").replace("}", "")          # _{E} → _E
-    s = re.sub(r"[^A-Za-z0-9_]+", "_", s)            # any leftover → _
-    s = re.sub(r"_+", "_", s).strip("_")             # tidy underscores
-    return s or "sym"
 
 
 def latex_to_semantic_graph(
